+++ tests/Core/Content/App/Manifest/Xml/MetadataTest.php
@@ -15,6 +15,7 @@ class MetadataTest extends TestCase
         static::assertEquals('SwagApp', $metaData->getName());
         static::assertEquals('shopware AG', $metaData->getAuthor());
         static::assertEquals('(c) by shopware AG', $metaData->getCopyright());
+        static::assertEquals('MIT', $metaData->getLicense());
         static::assertEquals('1.0.0', $metaData->getVersion());
         static::assertEquals('icon.png', $metaData->getIcon());